+++ include/ace-vm/TypeInfo.hpp
@@ -8,7 +8,7 @@ namespace vm {
 
 class TypeInfo {
 public:
-    TypeInfo(const char *name, int size, char **names);
+    TypeInfo(const char *name, size_t size, char **names);
     TypeInfo(const TypeInfo &other);
     TypeInfo &operator=(const TypeInfo &other);
     ~TypeInfo();
@@ -16,14 +16,14 @@ public:
     bool operator==(const TypeInfo &other) const;
 
     inline char *const GetName() const { return m_name; }
-    inline int GetSize() const { return m_size; }
+    inline size_t GetSize() const { return m_size; }
     inline char **const GetNames() const { return m_names; }
-    inline const char *GetMemberName(int index) const
+    inline const char *GetMemberName(size_t index) const
         { ASSERT(index < m_size); return m_names[index]; }
 
 private:
     char *m_name;
-    int m_size;
+    size_t m_size;
     char **m_names;
 };
